backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["linduix <abdullahhkhan3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.12.0"
aniso8601 = "9.0.1"
click = "8.1.3"
flask = "2.2.2"
flask-restful = "0.3.9"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.2"
pyjwt = "2.6.0"
pytz = "2022.7.1"
six = "1.16.0"
werkzeug = "2.2.2"
flask-sqlalchemy = "^3.0.2"
flask-jwt-extended = "^4.4.4"
sendgrid = "^6.9.7"
fastapi = {extras = ["all"], version = "^0.89.1"}
sqlalchemy = "^1.4.46"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
python-dotenv = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
